[Feat] server을 별도의 모듈로 분리
