+++ server/build.gradle
@@ -13,19 +13,6 @@ dependencies {
     testImplementation platform('org.junit:junit-bom:5.10.0')
     testImplementation 'org.junit.jupiter:junit-jupiter'
     implementation 'ch.qos.logback:logback-classic:1.5.6'
-    implementation project(':server')
-}
-
-jar {
-    manifest {
-        attributes(
-                'Main-Class': 'codesquad.Main'
-        )
-    }
-    from {
-        configurations.runtimeClasspath.collect { it.isDirectory() ? it : zipTree(it) }
-    }
-    duplicatesStrategy = DuplicatesStrategy.EXCLUDE
 }
 
 test {
