+++ server/src/main/java/server/connection/KeepAliveManager.java
@@ -1,0 +1,10 @@
+package server.connection;
+
+public interface KeepAliveManager {
+
+    boolean isAlive();
+
+    boolean isTimeout();
+
+    void incrementRequest();
+}
