+++ server/src/main/java/server/connection/OneTimeConnectionManager.java
@@ -1,0 +1,9 @@
+package server.connection;
+
+import java.net.Socket;
+
+public class OneTimeConnectionManager extends ConnectionManager {
+    public OneTimeConnectionManager(Socket socket) {
+        super(socket, new NoOpKeepAliveManager());
+    }
+}
