+++ server/src/main/java/server/exception/BadGrammarException.java
@@ -1,0 +1,4 @@
+package server.exception;
+
+public class BadGrammarException extends Exception {
+}
