+++ server/src/main/java/server/exception/ConnectedSocketException.java
@@ -1,0 +1,7 @@
+package server.exception;
+
+public class ConnectedSocketException extends ServerException {
+    public ConnectedSocketException(Throwable cause) {
+        super(cause);
+    }
+}
