+++ server/src/main/java/server/exception/ServerException.java
@@ -1,0 +1,7 @@
+package server.exception;
+
+public class ServerException extends RuntimeException {
+    public ServerException(Throwable cause) {
+        super(cause);
+    }
+}
