+++ server/src/main/java/server/http/model/body/Body.java
@@ -1,0 +1,13 @@
+package server.http.model.body;
+
+public class Body {
+    private final byte[] message;
+
+    public Body(byte[] message) {
+        this.message = message;
+    }
+
+    public byte[] getMessage() {
+        return message;
+    }
+}
