+++ server/src/main/java/server/http/model/startline/StatusCode.java
@@ -1,0 +1,25 @@
+package server.http.model.startline;
+
+public enum StatusCode {
+    OK(200, "OK"),
+    FOUND(302, "Found"),
+    BAD_REQUEST(400, "Bad Request"),
+    NOT_FOUND(404, "Not Found"),
+    INTERNAL_SERVER_ERROR(500, "Internal Server Error");
+
+    private final int code;
+    private final String message;
+
+    StatusCode(int code, String message) {
+        this.code = code;
+        this.message = message;
+    }
+
+    public int getCode() {
+        return code;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
