+++ server/src/main/java/server/processor/HttpRequestProcessor.java
@@ -1,0 +1,11 @@
+package server.processor;
+
+
+import server.http.model.HttpRequest;
+import server.http.model.HttpResponse;
+
+public interface HttpRequestProcessor {
+    HttpResponse process(HttpRequest request);
+
+    boolean supports(HttpRequest request);
+}
