+++ server/src/main/java/server/processor/StaticResourceProcessor.java
@@ -1,0 +1,4 @@
+package server.processor;
+
+public abstract class StaticResourceProcessor implements HttpRequestProcessor {
+}
