+++ src/main/java/codesquad/application/model/User.java
@@ -1,0 +1,22 @@
+package codesquad.application.model;
+
+public class User {
+    private String userId;
+    private String password;
+    private String nickname;
+
+    public User(String userId, String password, String nickname) {
+        this.userId = userId;
+        this.password = password;
+        this.nickname = nickname;
+    }
+
+    @Override
+    public String toString() {
+        return "User{" +
+                "userId='" + userId + '\'' +
+                ", password='" + password + '\'' +
+                ", nickname='" + nickname + '\'' +
+                '}';
+    }
+}
