--- src/main/java/codesquad/application/user/model/User.java
@@ -1,22 +1,0 @@
-package codesquad.application.user.model;
-
-public class User {
-    private String userId;
-    private String password;
-    private String nickname;
-
-    public User(String userId, String password, String nickname) {
-        this.userId = userId;
-        this.password = password;
-        this.nickname = nickname;
-    }
-
-    @Override
-    public String toString() {
-        return "User{" +
-                "userId='" + userId + '\'' +
-                ", password='" + password + '\'' +
-                ", nickname='" + nickname + '\'' +
-                '}';
-    }
-}
