--- src/main/java/codesquad/exception/BadGrammarException.java
@@ -1,4 +1,0 @@
-package codesquad.exception;
-
-public class BadGrammarException extends Exception {
-}
