--- src/main/java/codesquad/exception/ConnectedSocketException.java
@@ -1,7 +1,0 @@
-package codesquad.exception;
-
-public class ConnectedSocketException extends ServerException {
-    public ConnectedSocketException(Throwable cause) {
-        super(cause);
-    }
-}
