--- src/main/java/codesquad/exception/ServerException.java
@@ -1,7 +1,0 @@
-package codesquad.exception;
-
-public class ServerException extends RuntimeException {
-    public ServerException(Throwable cause) {
-        super(cause);
-    }
-}
