--- src/main/java/codesquad/http/model/body/Body.java
@@ -1,13 +1,0 @@
-package codesquad.http.model.body;
-
-public class Body {
-    private final byte[] body;
-
-    public Body(byte[] body) {
-        this.body = body;
-    }
-
-    public byte[] getBody() {
-        return body;
-    }
-}
