--- src/main/java/codesquad/http/model/header/Header.java
@@ -1,15 +1,0 @@
-package codesquad.http.model.header;
-
-public enum Header {
-    CONNECTION("connection"), CONTENT_TYPE("Content-Type"), CONTENT_LENGTH("Content-Length");
-
-    private final String fieldName;
-
-    Header(String fieldName) {
-        this.fieldName = fieldName;
-    }
-
-    public String getFieldName() {
-        return fieldName;
-    }
-}
