--- src/main/java/codesquad/http/model/startline/Target.java
@@ -1,20 +1,0 @@
-package codesquad.http.model.startline;
-
-public class Target {
-    private final String path;
-
-    public Target(String path) {
-        this.path = path;
-    }
-
-    public String getPath() {
-        return path;
-    }
-
-    @Override
-    public String toString() {
-        return "Target{" +
-                "path='" + path + '\'' +
-                '}';
-    }
-}
