--- src/main/java/codesquad/server/connection/KeepAliveManager.java
@@ -1,10 +1,0 @@
-package codesquad.server.connection;
-
-public interface KeepAliveManager {
-
-    boolean isAlive();
-
-    boolean isTimeout();
-
-    void incrementRequest();
-}
