--- src/main/java/codesquad/server/connection/OneTimeConnectionManager.java
@@ -1,9 +1,0 @@
-package codesquad.server.connection;
-
-import java.net.Socket;
-
-public class OneTimeConnectionManager extends ConnectionManager {
-    public OneTimeConnectionManager(Socket socket) {
-        super(socket, new NoOpKeepAliveManager());
-    }
-}
