--- src/main/java/codesquad/server/processor/HttpRequestProcessor.java
@@ -1,10 +1,0 @@
-package codesquad.server.processor;
-
-import codesquad.http.model.HttpRequest;
-import codesquad.http.model.HttpResponse;
-
-public interface HttpRequestProcessor {
-    HttpResponse process(HttpRequest request);
-
-    boolean supports(HttpRequest request);
-}
