--- src/main/java/codesquad/server/processor/StaticResourceProcessor.java
@@ -1,4 +1,0 @@
-package codesquad.server.processor;
-
-public abstract class StaticResourceProcessor implements HttpRequestProcessor {
-}
